Refactored rules/complexity test cases
This separates positive and negative test cases for rules. (#527)
