+++ detekt-rules/src/test/resources/cases/ComplexInterface.kt
@@ -2,33 +2,6 @@
 
 package cases
 
-interface TooLargeInterface {
-	fun f1()
-	fun f2()
-	val i1: Int
-	fun fImpl() { }
-}
-
-class ClassWithNestedInterface {
-
-	interface TooLargeNestedInterface {
-		fun f1()
-		fun f2()
-		val i1: Int
-		fun fImpl() { }
-	}
-}
-
-interface TooLargeInterfaceWithStaticDeclarations {
-	fun f1()
-
-	companion object {
-		fun sf() = 0
-		const val c = 0
-		val si = 0
-	}
-}
-
 interface InterfaceOk1 {
 	fun f1()
 	fun fImpl() {
