+++ detekt-rules/src/test/resources/cases/NestedClasses.kt
@@ -8,25 +8,9 @@ package cases
 @Suppress("unused")
 class NestedClasses {
 
-	override fun equals(other: Any?): Boolean {
-		return super.equals(other)
-	}
-
-	override fun hashCode(): Int {
-		return super.hashCode()
-	}
-
 	class InnerClass {
 
-		override fun equals(other: Any?): Boolean {
-			return super.equals(other)
-		}
-
 		class NestedInnerClass {
-
-			override fun hashCode(): Int {
-				return super.hashCode()
-			}
 
 			fun nestedLongMethod() {
 				if (true) {
